myapp/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dataclasses-json==0.5.8
dnspython==2.3.0
eventlet==0.33.3
firebase-admin==6.1.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.4
frozenlist==1.3.3
gevent==23.7.0
gevent-websocket==0.10.1
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
google-search-results==2.4.2
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.210
langchainplus-sdk==0.0.17
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
proto-plus==1.22.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
pyparsing==3.1.0
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
PyYAML==6.0
requests==2.31.0
rsa==4.9
simple-websocket==0.10.1
six==1.16.0
SQLAlchemy==2.0.16
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
